# services/api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API gateway for EXBanka microservices.",
        "title": "EXBanka API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/employees": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of all employees.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "List all employees",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.EmployeeListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new inactive employee and send an activation email.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Create employee",
                "parameters": [
                    {
                        "description": "New employee data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateEmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.employeeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/employees/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search employees by email, first name, last name, or position with pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Search employees",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by first name",
                        "name": "ime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by last name",
                        "name": "prezime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by position",
                        "name": "pozicija",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 20)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.EmployeeListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/employees/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a single employee by their numeric ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Get employee by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.employeeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update all fields of an existing employee.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "employees"
                ],
                "summary": "Update employee",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Employee ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated employee data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateEmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.employeeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateEmployeeRequest": {
            "type": "object",
            "required": [
                "address",
                "date_of_birth",
                "department",
                "email",
                "first_name",
                "gender",
                "last_name",
                "phone_number",
                "position",
                "username"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Bulevar Kralja Aleksandra 73"
                },
                "date_of_birth": {
                    "type": "string",
                    "example": "1990-01-15"
                },
                "department": {
                    "type": "string",
                    "example": "Retail"
                },
                "email": {
                    "type": "string",
                    "example": "marko@exbanka.rs"
                },
                "first_name": {
                    "type": "string",
                    "example": "Marko"
                },
                "gender": {
                    "type": "string",
                    "example": "M"
                },
                "last_name": {
                    "type": "string",
                    "example": "Marković"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+381641234567"
                },
                "position": {
                    "type": "string",
                    "example": "Teller"
                },
                "username": {
                    "type": "string",
                    "example": "mmarkovic"
                }
            }
        },
        "handlers.EmployeeListResponse": {
            "type": "object",
            "properties": {
                "employees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.employeeResponse"
                    }
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.UpdateEmployeeRequest": {
            "type": "object",
            "required": [
                "address",
                "date_of_birth",
                "department",
                "email",
                "first_name",
                "gender",
                "last_name",
                "phone_number",
                "position",
                "username"
            ],
            "properties": {
                "active": {
                    "type": "boolean",
                    "example": true
                },
                "address": {
                    "type": "string",
                    "example": "Bulevar Kralja Aleksandra 73"
                },
                "date_of_birth": {
                    "type": "string",
                    "example": "1990-01-15"
                },
                "department": {
                    "type": "string",
                    "example": "Retail"
                },
                "email": {
                    "type": "string",
                    "example": "marko@exbanka.rs"
                },
                "first_name": {
                    "type": "string",
                    "example": "Marko"
                },
                "gender": {
                    "type": "string",
                    "example": "M"
                },
                "last_name": {
                    "type": "string",
                    "example": "Marković"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "LOANS"
                    ]
                },
                "phone_number": {
                    "type": "string",
                    "example": "+381641234567"
                },
                "position": {
                    "type": "string",
                    "example": "Teller"
                },
                "username": {
                    "type": "string",
                    "example": "mmarkovic"
                }
            }
        },
        "handlers.employeeResponse": {
            "type": "object",
            "properties": {
                "adresa": {
                    "type": "string"
                },
                "aktivan": {
                    "type": "boolean"
                },
                "broj_telefona": {
                    "type": "string"
                },
                "datum_rodjenja": {
                    "type": "string"
                },
                "departman": {
                    "type": "string"
                },
                "dozvole": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ime": {
                    "type": "string"
                },
                "pol": {
                    "type": "string"
                },
                "pozicija": {
                    "type": "string"
                },
                "prezime": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}